+++ CHANGELOG.md
@@ -4,3 +4,5 @@ Please see each crate's change log below:
 
 - [`data_privacy`](./crates/data_privacy/CHANGELOG.md)
 - [`data_privacy_macros`](./crates/data_privacy_macros/CHANGELOG.md)
+- [`fundle`](./crates/fundle/CHANGELOG.md)
+- [`fundle_macros`](./crates/fundle_macros/CHANGELOG.md)